requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Pillow==8.2.0
streamlit==0.80.0
torch==1.8.1+cpu 
torchvision==0.9.1+cpu